refactor: moved button component from notes to global components
