+++ src/renderer/src/lib/notes/addnoteForm.svelte
@@ -2,7 +2,7 @@
     import NotesStore from "./stores/NotesStore";
     import type { Note } from "../../types/note";
     import { createEventDispatcher } from "svelte";
-    import Button from "./Button.svelte";
+    import Button from "../global-components/Button.svelte";
     import EmojiSelector from 'svelte-emoji-selector';
 
     let dispatch = createEventDispatcher();
